Move adapter specific tests to the adapter folder
Since #46865, sqlite3 specific tests were running for all adapter.

That happened because the `current_adapter?(:SQLite3Adapter)` check was
moved to the test case class but not for all of them.

To avoid this issue to happen again, just move those tests to each
adapter.
